backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
motor>=3.6.0
pymongo>=4.9.0
python-dotenv>=1.0.1
pydantic-settings>=2.6.0

langchain>=0.3.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.4
langchain-text-splitters>=0.3.0
langchain-google-genai
chromadb>=0.5.0
pypdf>=5.0.0
